// src/options/app/lib/type.ts
export type OptionUI_PageID = string
export type OptionUI_CategoryID = string
export type OptionUI_ItemID = string

export type OptionUI_Anchors = Array<OptionUI_Anchor>
export interface OptionUI_Anchor {
    id: string
    level: number
    label: string
    target: {
        page: string
        category: string
        id: string
    }
}

/**
 * オプションページ上の説明文クラス
 * @param {string} text - メイン文章
 * @param {string} small - サブ文章
 * @param {string} attention - 警告文（赤字）
 * @param {string} hidden - 非表示
 * @param {Array<string>} keywords - 絵kん咲くようキーワード
 */
export class OptionUI_Descriptions {
    attention?: string
    text?: string
    small?: string
    hidden?: string
    keywords?: Array<string>
}
// ...
export interface OptionUI_Page {
    id: OptionUI_PageID
    type?: "page" | "divider" | "header"
    title?: string
    description?: string
    icon?: OptionUI_FontAwesomeIcon
    targetUrl?: Array<string>
    defaultCategory?: OptionUI_CategoryID
    categories?: Array<OptionUI_Category> 
    parent?: OptionUI_PageID
    position?: OptionUI_Page_Position
    noindex?: boolean
    hideToc?: "hide"|"empty"|"show"
    popup?: boolean
}

export type OptionUI_Page_Position = "top" | "bottom" | "hide"

export interface OptionUI_FontAwesomeIcon {
    icon: string
    prefix?: string
    variant?: string
}

/************************************************************************************* */
/**
 * オプションページのカテゴリ
 * @param {OptionUI_CategoryID} id - カテゴリID
 * @param {string} title - 表示名
 * @param {OptionUI_Descriptions} description - 説明文
 */
export interface OptionUI_Category {
    id: OptionUI_CategoryID
    title: string
    description?: OptionUI_Descriptions
    hideTitle?: boolean
    hideDescription?: boolean
}